statistic: mean is a float value (again)

+++ include/rpiraw.h
@@ -58,7 +58,7 @@
                                    const unsigned npixs);
 
     float rpiraw_calc_acm_by_hist(uint32_t *hist, const unsigned len,
-                                  const unsigned npixs, const uint32_t mean);
+                                  const unsigned npixs, const float mean);
 
     int rpiraw_histogram_equalize_rgb888(uint8_t *dst, const unsigned ld_dst,
                                          uint8_t *src, const unsigned ld_src,
